+++ build-plugin/CMakeFiles/SappChoirPlugin.dir/cmake_clean_target.cmake
@@ -1,3 +1,0 @@
-file(REMOVE_RECURSE
-  "SappChoirPlugin_artefacts/Release/libSappChoir_SharedCode.a"
-)